+++ Runtime/Logging/LoogaLogLevel.cs
@@ -1,0 +1,13 @@
+namespace LoogaSoft.Logging
+{
+    public enum LoogaLogLevel
+    {
+        Trace = 0,
+        Debug = 1,
+        Info = 2,
+        Warning = 3,
+        Error = 4,
+        Exception = 5,
+        Off = 6
+    }
+}